Refactor include guards and add missing licenses
