+++ src/Accessories/Turnouts.h
@@ -17,8 +17,8 @@
  *  along with CommandStation.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-#ifndef COMMANDSTATION_ACCESSORIES_TURNOUTS_H_
-#define COMMANDSTATION_ACCESSORIES_TURNOUTS_H_
+#ifndef ACCESSORIES_TURNOUTS_H_
+#define ACCESSORIES_TURNOUTS_H_
 
 #include <Arduino.h>
 #include "../DCC/DCC.h"
@@ -44,4 +44,4 @@ struct Turnout{
   static void show(Print* stream, int=0);
 };
   
-#endif  // COMMANDSTATION_ACCESSORIES_TURNOUTS_H_
+#endif  // ACCESSORIES_TURNOUTS_H_
