+++ src/Utils/FreeMemory.h
@@ -1,5 +1,35 @@
-#ifndef COMMANDSTATION_DCC_FREEMEMORY_H_
-#define COMMANDSTATION_DCC_FREEMEMORY_H_
+/*
+ *  FreeMemory.h
+ * 
+ *  This file is part of CommandStation, which is licensed under GPLv3.
+ *  FreeMemory.h is from a work of Michael P. Flaga, and is under the MIT
+ *  License.
+ *
+ *  The MIT License (MIT)
+ *
+ *  Copyright (c) 2012 Michael P. Flaga
+
+ *  Permission is hereby granted, free of charge, to any person obtaining a copy
+ *  of this software and associated documentation files (the "Software"), to deal
+ *  in the Software without restriction, including without limitation the rights
+ *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ *  copies of the Software, and to permit persons to whom the Software is
+ *  furnished to do so, subject to the following conditions:
+
+ *  The above copyright notice and this permission notice shall be included in all
+ *  copies or substantial portions of the Software.
+
+ *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ *  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ *  SOFTWARE.
+ */
+
+#ifndef UTILS_FREEMEMORY_H_
+#define UTILS_FREEMEMORY_H_
 
 // thanks go to  https://github.com/mpflaga/Arduino-MemoryFree
 #ifdef __arm__
@@ -17,4 +47,4 @@ int freeMemory() {
   return __brkval ? &top - __brkval : &top - __malloc_heap_start;
 #endif  // __arm__
 }
-#endif  // COMMANDSTATION_DCC_FREEMEMORY_H_
+#endif  // UTILS_FREEMEMORY_H_
